refac: Segregate ProcessStepResultRepository from ProcessOutputRepository
