+++ source/dotnet/Services/Tests/Infrastructure/Processes/ProcessStepResultRepositoryTests.cs
@@ -1,0 +1,239 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using AutoFixture.Xunit2;
+using Azure;
+using Azure.Storage.Files.DataLake;
+using Azure.Storage.Files.DataLake.Models;
+using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
+using Energinet.DataHub.Wholesale.Application.ProcessResult;
+using Energinet.DataHub.Wholesale.Contracts;
+using Energinet.DataHub.Wholesale.Domain.GridAreaAggregate;
+using Energinet.DataHub.Wholesale.Domain.ProcessStepResultAggregate;
+using Energinet.DataHub.Wholesale.Infrastructure.Processes;
+using Energinet.DataHub.Wholesale.Tests.Infrastructure.BasisData;
+using FluentAssertions;
+using Moq;
+using Xunit;
+using Xunit.Categories;
+
+namespace Energinet.DataHub.Wholesale.Tests.Infrastructure.Processes;
+
+[UnitTest]
+public class ProcessStepResultRepositoryTests
+{
+    [Theory]
+    [AutoMoqData]
+    public async Task GetAsync_ReturnsProcessActorResult(
+        [Frozen] Mock<IProcessResultPointFactory> processResultFactoryMock,
+        [Frozen] Mock<DataLakeFileSystemClient> dataLakeFileSystemClientMock,
+        [Frozen] Mock<DataLakeDirectoryClient> dataLakeDirectoryClientMock,
+        [Frozen] Mock<DataLakeFileClient> dataLakeFileClientMock,
+        [Frozen] Mock<Response<bool>> responseMock)
+    {
+        // Arrange
+        const string pathWithKnownExtension = "my_file.json";
+        var asyncPageable = CreateAsyncPageableWithOnePathItem(pathWithKnownExtension);
+        var stream = new Mock<Stream>();
+
+        dataLakeDirectoryClientMock
+            .Setup(client => client.GetPathsAsync(false, false, It.IsAny<CancellationToken>()))
+            .Returns(asyncPageable);
+        responseMock.Setup(res => res.Value).Returns(true);
+        dataLakeDirectoryClientMock.Setup(dirClient => dirClient.ExistsAsync(default))
+            .ReturnsAsync(responseMock.Object);
+        dataLakeFileSystemClientMock.Setup(x => x.GetDirectoryClient(It.IsAny<string>()))
+            .Returns(dataLakeDirectoryClientMock.Object);
+        dataLakeFileSystemClientMock.Setup(x => x.GetFileClient(pathWithKnownExtension))
+            .Returns(dataLakeFileClientMock.Object);
+        dataLakeFileClientMock
+            .Setup(x => x.OpenReadAsync(It.IsAny<bool>(), It.IsAny<long>(), It.IsAny<int?>(), default))
+            .ReturnsAsync(stream.Object);
+        var processResultPoint = new ProcessResultPoint("1.00", "A04", "2022-05-31T22:00:00");
+        processResultFactoryMock.Setup(x => x.GetPointsFromJsonStreamAsync(stream.Object))
+            .ReturnsAsync(new List<ProcessResultPoint>
+            {
+                processResultPoint,
+            });
+
+        var sut = new ProcessStepResultRepository(
+            dataLakeFileSystemClientMock.Object,
+            processResultFactoryMock.Object);
+
+        // Act
+        var actual = await sut.GetAsync(Guid.NewGuid(), new GridAreaCode("123"));
+
+        // Assert
+        actual.Should().NotBeNull();
+    }
+
+    [Theory]
+    [AutoMoqData]
+    public async Task GetResultFileStreamAsync_WhenDirectoryDoesNotExist_ThrowsException(
+        [Frozen] Mock<IProcessResultPointFactory> processResultFactoryMock,
+        [Frozen] Mock<DataLakeFileSystemClient> dataLakeFileSystemClientMock,
+        [Frozen] Mock<DataLakeDirectoryClient> dataLakeDirectoryClientMock,
+        [Frozen] Mock<Response<bool>> responseMock)
+    {
+        // Arrange
+        var asyncPageable = CreateAsyncPageableWithOnePathItem("my_file.json");
+
+        dataLakeDirectoryClientMock
+            .Setup(client => client.GetPathsAsync(false, false, It.IsAny<CancellationToken>()))
+            .Returns(asyncPageable);
+        dataLakeFileSystemClientMock.Setup(x => x.GetDirectoryClient(It.IsAny<string>()))
+            .Returns(dataLakeDirectoryClientMock.Object);
+        dataLakeFileSystemClientMock.Setup(x => x.GetFileClient(It.IsAny<string>()))
+            .Returns((Func<DataLakeFileClient>)null!);
+        dataLakeDirectoryClientMock.Setup(dirClient => dirClient.ExistsAsync(default))
+            .ReturnsAsync(responseMock.Object);
+        responseMock.Setup(res => res.Value).Returns(true);
+
+        var sut = new ProcessStepResultRepository(
+            dataLakeFileSystemClientMock.Object,
+            processResultFactoryMock.Object);
+
+        // Act and Assert
+        await sut
+            .Invoking(s => s.GetAsync(Guid.NewGuid(), new GridAreaCode("123")))
+            .Should()
+            .ThrowAsync<InvalidOperationException>();
+    }
+
+    [Theory]
+    [AutoMoqData]
+    public async Task GetResultFileStreamAsync_WhenNoFileClientFound_ThrowsException(
+        [Frozen] Mock<IProcessResultPointFactory> processResultFactoryMock,
+        [Frozen] Mock<DataLakeFileSystemClient> dataLakeFileSystemClientMock,
+        [Frozen] Mock<DataLakeDirectoryClient> dataLakeDirectoryClientMock,
+        [Frozen] Mock<Response<bool>> responseMock)
+    {
+        // Arrange
+        responseMock.Setup(res => res.Value).Returns(true);
+        dataLakeFileSystemClientMock.Setup(x => x.GetDirectoryClient(It.IsAny<string>()))
+            .Returns(dataLakeDirectoryClientMock.Object);
+        dataLakeDirectoryClientMock.Setup(dirClient => dirClient.ExistsAsync(default))
+            .ReturnsAsync(responseMock.Object);
+
+        var sut = new ProcessStepResultRepository(
+            dataLakeFileSystemClientMock.Object,
+            processResultFactoryMock.Object);
+
+        // Act and Assert
+        await sut
+            .Invoking(s => s.GetAsync(Guid.NewGuid(), new GridAreaCode("123")))
+            .Should()
+            .ThrowAsync<Exception>();
+    }
+
+    [Theory]
+    [AutoMoqData]
+    public async Task GetResultFileStreamAsync_WhenFileExtensionNotFound_ThrowException(
+        [Frozen] Mock<IProcessResultPointFactory> processResultFactoryMock,
+        [Frozen] Mock<DataLakeFileSystemClient> dataLakeFileSystemClientMock,
+        [Frozen] Mock<DataLakeDirectoryClient> dataLakeDirectoryClientMock,
+        [Frozen] Mock<Response<bool>> responseMock)
+    {
+        // Arrange
+        const string pathWithUnknownExtension = "my_file.xxx";
+        var pathItem = DataLakeModelFactory.PathItem(
+            pathWithUnknownExtension,
+            false,
+            DateTimeOffset.Now,
+            ETag.All,
+            1,
+            "owner",
+            "group",
+            "permissions");
+        var page = Page<PathItem>.FromValues(new[] { pathItem }, null, Mock.Of<Response>());
+        var asyncPageable = AsyncPageable<PathItem>.FromPages(new[] { page });
+
+        dataLakeDirectoryClientMock
+            .Setup(client => client.GetPathsAsync(false, false, It.IsAny<CancellationToken>()))
+            .Returns(asyncPageable);
+        responseMock.Setup(res => res.Value).Returns(true);
+        dataLakeFileSystemClientMock.Setup(x => x.GetDirectoryClient(It.IsAny<string>()))
+            .Returns(dataLakeDirectoryClientMock.Object);
+
+        dataLakeDirectoryClientMock.Setup(dirClient => dirClient.ExistsAsync(default))
+            .ReturnsAsync(responseMock.Object);
+
+        var sut = new ProcessStepResultRepository(
+            dataLakeFileSystemClientMock.Object,
+            processResultFactoryMock.Object);
+
+        // Act and Assert
+        await sut
+            .Invoking(s => s.GetAsync(Guid.NewGuid(), new GridAreaCode("123")))
+            .Should()
+            .ThrowAsync<Exception>();
+    }
+
+    [Fact]
+    public static async Task GetResultFileSpecification_MatchesContract()
+    {
+        // Arrange
+        const string batchId = "eac4a18d-ed5f-46ba-bfe7-435ec0323519";
+        const string gridAreaCode = "123";
+        var calculationFilePathsContract = await CalculationFilePathsContract.GetAsync();
+        var expected = calculationFilePathsContract.ResultFile;
+
+        // Act
+        var actual = ProcessStepResultRepository.GetResultFileSpecification(new Guid(batchId), new GridAreaCode(gridAreaCode));
+
+        // Assert
+        actual.Extension.Should().Be(expected.Extension);
+        actual.Directory.Should().MatchRegex(expected.DirectoryExpression);
+    }
+
+    [Theory]
+    [AutoMoqData]
+    public async Task GetResultAsync_TimeSeriesPoint_IsRead(
+        [Frozen] Mock<IProcessStepResultRepository> processActorResultRepositoryMock)
+    {
+        // Arrange
+        var time = new DateTimeOffset(2022, 05, 15, 22, 15, 0, TimeSpan.Zero);
+        var quantity = 1.000m;
+        var quality = "A04";
+
+        const string gridAreaCode = "805";
+        var batchId = Guid.NewGuid();
+
+        var sut = new ProcessStepResultApplicationService(processActorResultRepositoryMock.Object, new ProcessStepResultMapper());
+
+        processActorResultRepositoryMock.Setup(p => p.GetAsync(batchId, new GridAreaCode(gridAreaCode)))
+            .ReturnsAsync(new ProcessStepResult(new[] { new TimeSeriesPoint(time, quantity, quality) }));
+
+        // Act
+        var actual = await sut.GetResultAsync(
+            new ProcessStepResultRequestDto(
+                batchId,
+                gridAreaCode,
+                ProcessStepType.AggregateProductionPerGridArea));
+
+        // Assert
+        actual.TimeSeriesPoints.First().Time.Should().Be(time);
+        actual.TimeSeriesPoints.First().Quantity.Should().Be(quantity);
+        actual.TimeSeriesPoints.First().Quality.Should().Be(quality);
+    }
+
+    private static AsyncPageable<PathItem> CreateAsyncPageableWithOnePathItem(string path)
+    {
+        var pathItem = DataLakeModelFactory
+            .PathItem(path, false, DateTimeOffset.Now, ETag.All, 1, "owner", "group", "permissions");
+        var page = Page<PathItem>.FromValues(new[] { pathItem }, null, Mock.Of<Response>());
+        var asyncPageable = AsyncPageable<PathItem>.FromPages(new[] { page });
+        return asyncPageable;
+    }
+}
